backend/requirements.txt
fastapi~=0.115.3
google-generative-ai
python-dotenv~=1.0.1
pydantic~=2.9.2